+++ Python_OOP/Classes_Objects/Exercise/project_05/task.py
